requirements.txt
python-telegram-bot>=22.0
anthropic>=0.40.0
starlette>=0.27.0
uvicorn>=0.23.0
